sign up modal and map fixing
