+++ frontend/src/app/App.tsx
@@ -1,76 +1,15 @@
-import React, { useCallback, useMemo, useRef, useState } from 'react';
+import React from 'react';
 import { SafeAreaView, StyleSheet, View } from 'react-native';
-import SearchBar from '../components/SearchBar';
-import Header from '../components/Header';
-import Logo from '../components/Logo';
-import Map from '../components/Map';
-import FiltersDropdown from '../components/FiltersDropdown';
-import { GestureHandlerRootView } from 'react-native-gesture-handler';
-import BottomSheet, { BottomSheetFlatList } from '@gorhom/bottom-sheet';
-import ShelterInfoPanel from '../components/ShelterInfoPanel';
-import { Shelter, shelters } from '../sheltersTest';
+import CompleteMap from '../components/CompleteMap';
 import SignUpModal from '../components/SignUpModal';
 
 export const App = () => {
-  const sheetRef = useRef<BottomSheet>(null);
-  const snapPoints = useMemo(() => ['15%', '60%', '90%'], []);
-  const [selectedShelter, setSelectedShelter] = useState<Shelter | null>(null);
-
-  const handleMarkerPress = useCallback((shelter: Shelter) => {
-    setSelectedShelter(shelter);
-    sheetRef.current?.snapToIndex(1);
-  }, []);
-
-  const renderItem = useCallback(
-    ({ item }: { item: Shelter }) => (
-      <ShelterInfoPanel
-        title={item.title}
-        description={item.description}
-        style={styles.itemContainer}
-      />
-    ),
-    []
-  );
-
   return (
     <SafeAreaView style={styles.safeArea}>
-      <View style={styles.logoContainer}>
-        <Logo />
-      </View>
-      <View style={styles.centeredView}>
-        <SignUpModal />
-      </View>
-      <View style={styles.searchBarContainer}>
-        <SearchBar />
-      </View>
-      <View style={styles.headerContainer}>
-        <Header />
-      </View>
-      <View style={styles.filtersDropdownContainer}>
-        <FiltersDropdown />
-      </View>
-      <GestureHandlerRootView style={{ flex: 1 }}>
-        <Map onMarkerPress={handleMarkerPress} style={styles.map} />
-        <BottomSheet
-          ref={sheetRef}
-          snapPoints={snapPoints}
-          style={styles.bottomSheet}
-        >
-          {selectedShelter ? (
-            <ShelterInfoPanel
-              title={selectedShelter.title}
-              description={selectedShelter.description}
-              style={styles.itemContainer}
-            />
-          ) : (
-            <BottomSheetFlatList
-              data={shelters}
-              keyExtractor={(item) => item.id.toString()}
-              renderItem={renderItem}
-            />
-          )}
-        </BottomSheet>
-      </GestureHandlerRootView>
+      <CompleteMap />
+      {/*<View style={styles.centeredView}>*/}
+      {/*  <SignUpModal />*/}
+      {/*</View>*/}
     </SafeAreaView>
   );
 };
@@ -79,43 +18,6 @@ const styles = StyleSheet.create({
   safeArea: {
     flex: 1,
     backgroundColor: '#FFFFFF',
-  },
-  container: {
-    flex: 1,
-  },
-  itemContainer: {
-    marginHorizontal: 29,
-    marginTop: 29,
-  },
-  logoContainer: {
-    alignItems: 'flex-start',
-    paddingHorizontal: 20,
-    paddingTop: 11,
-    paddingBottom: 18,
-  },
-  searchBarContainer: {
-    alignItems: 'center',
-    paddingHorizontal: 13,
-    paddingBottom: 22,
-  },
-  headerContainer: {
-    alignItems: 'center',
-    paddingHorizontal: 25,
-    paddingBottom: 30,
-  },
-  filtersDropdownContainer: {
-    alignItems: 'flex-start',
-    paddingHorizontal: 13,
-    paddingBottom: 12,
-  },
-  map: {
-    width: '100%',
-    height: '100%',
-  },
-  bottomSheet: {
-    borderTopLeftRadius: 33,
-    borderTopRightRadius: 33,
-    alignItems: 'center',
   },
   centeredView: {
     flex: 1,
